+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Open commit & push dialog after "done"
 - Set `Co-authored-by:` trailer to commit message automatically
+- Support start Zoom screenshare on Windows
 
 ## 1.0-beta2
 - Add shortcut for show done dialog (ALT+M, D)